move getBuyPrice/getSellPrice to shop model
