+++ app/rooms/commands/game-commands.ts
@@ -16,6 +16,7 @@ import Player from "../../models/colyseus-models/player"
 import { isOnBench } from "../../models/colyseus-models/pokemon"
 import PokemonFactory from "../../models/pokemon-factory"
 import { PVEStages } from "../../models/pve-stages"
+import { getBuyPrice, getSellPrice } from "../../models/shop"
 import { getAvatarString } from "../../public/src/utils"
 import {
   IClient,
@@ -96,7 +97,7 @@ export class OnShopCommand extends Command<
       pokemon.evolutionRule instanceof CountEvolutionRule &&
       pokemon.evolutionRule.canEvolveIfBuyingOne(pokemon, player)
 
-    let cost = PokemonFactory.getBuyPrice(name, this.state.specialGameRule)
+    let cost = getBuyPrice(name, this.state.specialGameRule)
     const freeSpaceOnBench = getFreeSpaceOnBench(player.board)
     const hasSpaceOnBench = freeSpaceOnBench > 0 || isEvolution
 
@@ -152,7 +153,7 @@ export class OnRemoveFromShopCommand extends Command<
       return
 
     const name = player.shop[index]
-    const cost = PokemonFactory.getBuyPrice(name, this.state.specialGameRule)
+    const cost = getBuyPrice(name, this.state.specialGameRule)
     if (player.money >= cost) {
       player.shop = player.shop.with(index, Pkm.DEFAULT)
       this.state.shop.releasePokemon(name, player)
@@ -521,7 +522,7 @@ export class OnSellDropCommand extends Command<
 
       if (pokemon) {
         this.state.shop.releasePokemon(pokemon.name, player)
-        player.money += PokemonFactory.getSellPrice(
+        player.money += getSellPrice(
           pokemon.name,
           pokemon.shiny,
           this.state.specialGameRule
